Add files via upload
This code imports the "re" and "Ast" modules. The re (regular expression) module provides functions for working with regular expressions, which are a way to match patterns in strings. The "Ast" (abstract syntax tree) module provides functions for parsing and interpreting Python code.

The code defines two regular expressions: "regular_expression" and "regular_expression_2". These regular expressions contain characters and operators for matching strings.

The code then uses the findall function from the re module to search for patterns in the regular expressions. The "findall" function returns a list of all the substrings that match the given pattern. The pattern in this case is defined using a regular expression: r"[.\()\*\w+]". This regular expression matches any character that is a period (.), left or right parentheses, an asterisk (*), or a word character (\w).

The code then uses the "ast.parse "function to parse the string "print(3+(5*2))" and create an abstract syntax tree (AST) for it. The AST is a tree-like structure that represents the structure of the code. The code then prints the AST using the "ast.dump" function and executes the code using the compile and exec functions.

The code does the same thing for the string "print('a+(b*c)')", creating an AST and executing the code.

Finally, the code prints the lists returned by the "findall" function for the two regular expressions.

The output of this code should be the AST for the two strings, the result of executing the two strings (which is printing the strings to the console), and the lists of substrings that match the pattern in the regular expressions.